Add time_unittest to FTL, actually build source files
This adds a trivial unit test for ftl's time routine. Since time is
implemented in source files and not entirely in headers like array
this also requires adding ftl's source files to the build. This
configures ftl as a static_library as it's expected to contain some
amount of unused things that don't need to be linked into the
consumer if they are not used.

Change-Id: If90467188bb0667e51be047bb0780cadb4bd007c