Enable multisampling.
Introduces new SSDO artifacts at the edges of shadow-casters:
sometimes the samples of the shadow-castee are unshadowed.
This is particularly noticable when the caster is dark-colored
and the castee is light-colored.

Finally, this change allows side-by-side performance comparison
between Linux and Fuchsia on the Acer Switch 12 Alpha (previous
hacks to run on Fuchsia broke Linux on that particular device).

Change-Id: Iacbaee8be59e32bb07e3f28c2c7e41151cb25af6