Accelerate SSDO via a lookup table of known-to-be unshadowed areas.
The basic idea is to start with an 8x-downsampled depth buffer,
and to create a lookup table of the same size (actually, the lookup
table is 4x smaller in each dimension: we pack a 4x4 region into the
bits of a single RGBA pixel).

We use a GPU compute kernel to look for depth discontinuities, in order
to determine when a table-cell contains geometry that may cast a shadow
upon a left/right/up/down-ward neighbor.  Then, because the maximum
shadow radius is greater than 8 pixels, we extend the potentially-shadowed
region in each direction.

Change-Id: Iec7c7a414445c5f151e5c495ace9b9f4b0d89cd7