Add Reffable, TypedReffable, Ownable and Owner classes.
The first is a non-threadsafe ref-counted base class that
is suitable for use with ftl::RefPtr.  It provides a
virtual OnZeroRefCount() method that subclasses can use to
avoid immediate destruction when the ref-count hits zero.

TypedReffable provides a framework for dynamic type-checking.

Ownable inherits from TypedReffable, and returns itself to
its Owner (if any) when its ref-count reaches zero.  The
Owner is responsible for deciding whether to destroy the
object, recycle it, etc.

Change-Id: I1ec5ad3c7c4cd48351cc5449b3d42b51af5125f6