Guarantee that ResourceRecyclers are able to recycle.
More specifically, guarantee that they are properly
hooked up to a CommandBufferSequenceListener, so
that they know when it is safe to recycle.  Before
this CL, it is possible for a ResourceRecycler to
keep resources in limbo forever.

For example, this manifested as an error upon shutdown.
("Check failed: unused_resources_.empty().")

Change-Id: Iacd29e1d964d16df33840366b3e7c1eb963ad131