Enhanced clipping support.
Previously, an escher::Object could clip a list of
children, but with some restrictions that are now
relaxed.

First, the children could only be clipped by the Object
itself (or rather, its shape).  Now, an Object has an
additional list of "clipping children" whose shapes are
also written into the stencil buffer at the same time
as the parent.  As a result, the other list of "clipped
children" is clipped both by the parent and the "clipping
children".

Second, it is now legal to specify an Object with a null
Material.  If this object is a clipper, then it avoids
writing to the color attachment (if it isn't a clipper,
then it is simply omitted from the ModelDisplayList).

Change-Id: Idcc5196ba799c718503e6d21c280d19b11973cfa